feat(Tester): enable convertion type on Point2
feat(Tester): use SetPixel instead of LineTo for drawing a pixel and add deleteDc in DrawLine

fix(Tester): correct calculation of the rect of a sprite

+++ Tester/Own-rendering-windows/include/Rect.hpp
@@ -6,7 +6,7 @@ namespace tester
 {
     struct Rect
     {
-        Point2<uint32_t> pos;
-        Point2<uint32_t> size;
+        Point2<float> pos;
+        Point2<float> size;
     };
 }
